chore: rename ObjectCache directory to caches
